Change how battles are created
This was motivated by the script crashing if someone was using a default profile picture. Now in every case where it's possible to determine which party is associated with which player, it should do so. It should also crash less often in irregular cases such as someone playing in two battles at once.