Bug fixes and quality of life changes
Fixed an issue where the program couldn't determine who used which move if both players used the same character and a hit was resisted.
Fixed an issue that caused the program to not come up with any move sequences when the character had Dominate.
Fixed an issue where the program deduced a player's ID incorrectly if they were using a server avatar and their avatar ID started with a number.
Now print the current character's buffs, debuffs, and statuses.
Delete battles where a character is undefined.