Add websocket to program
This will make bidirectional communication between the program and the webpage client, with the server acting as an intermediary, much easier. This will make future planned features such as a webpage button to add a rune to a player's party or prompting the webpage user to determine which player used what move.